+++ src/dedalus_labs/types/shared/__init__.py
@@ -2,3 +2,6 @@
 
 from .dedalus_model import DedalusModel as DedalusModel
 from .dedalus_model_choice import DedalusModelChoice as DedalusModelChoice
+from .response_format_text import ResponseFormatText as ResponseFormatText
+from .response_format_json_object import ResponseFormatJSONObject as ResponseFormatJSONObject
+from .response_format_json_schema import ResponseFormatJSONSchema as ResponseFormatJSONSchema